:white_check_mark: test(frontend): 迁移 validate 测试到独立文件，补充断言消息
