+++ src/frontend/tests/mod.rs
@@ -3,3 +3,4 @@
 //! 包含前端编译器各模块的测试
 
 mod config;
+mod validate;
