+++ src/frontend/validate.rs
@@ -91,47 +91,3 @@ pub fn validate_source(source: &str) -> ValidateResult {
     cache.insert(hash, Arc::new(result.clone()));
     result
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_empty_input() {
-        let result = validate_source("");
-        assert!(result.diagnostics.is_empty(), "空输入不应产生诊断");
-        assert!(result.module.is_some(), "空输入应产生模块 AST");
-    }
-
-    #[test]
-    fn test_undeclared_let() {
-        // `let` 作为未声明的标识符使用 —— 解析成功但类型检查报错
-        let result = validate_source("let x = 1");
-        assert!(!result.diagnostics.is_empty(), "未声明的 `let` 应产生诊断");
-        assert!(result.module.is_some(), "解析成功时模块 AST 应存在");
-    }
-
-    #[test]
-    fn test_valid_assignment() {
-        let result = validate_source("x = 1");
-        assert!(result.diagnostics.is_empty(), "合法赋值不应产生诊断");
-        assert!(result.module.is_some(), "合法赋值应产生模块 AST");
-    }
-
-    #[test]
-    fn test_cache_hit() {
-        // 连续两次调用相同源码应返回相同结果（内容等价 + 缓存命中）
-        let a = validate_source("x = 1");
-        let b = validate_source("x = 1");
-        assert_eq!(a.diagnostics.len(), b.diagnostics.len());
-        assert_eq!(a.module.is_some(), b.module.is_some());
-    }
-
-    #[test]
-    fn test_syntax_error() {
-        // `@` 不能作为语句开始 —— 显式产生语法错误
-        let result = validate_source("@");
-        assert!(!result.diagnostics.is_empty(), "语法错误应产生诊断");
-        assert!(result.module.is_none(), "语法错误时模块 AST 应为 None");
-    }
-}
